ME-36 Создание сцены для разработки UI контента для планшета
Dev:

- Созданы сцены `CreateTabletRoomUITemplate` и `CreateTabletRoomUI`. `CreateTabletRoomUITemplate` это шаблон сцены, что бы после очередного разработанного UI не пришлось очищать созданный UI. А `CreateTabletRoomUI` это сцена, в которой производится разработка UI
- Необходимые скрипты вынесены в SDK
