+++ Assets/ApplicationContent/APIScripts/External/Tablet/CreateTabletRoomUI.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a020561d4a6679646bc8fe77e866250c
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
