+++ Assets/ApplicationContent/APIScripts/External/Tablet/CreateTabletRoomUITemplate.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f1a5f5af721902d448d8cc14774e800d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
