+++ Assets/ApplicationContent/APIScripts/External/Tablet/RoomContentTablet.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b42687060fff8f546ab08a3d4c9462e9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
