+++ Assets/ApplicationContent/APIScripts/External/Tablet/UI/UITabletRoomContentSupplier.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+
+/// <summary>
+/// <para>Компонент передающий планшету контент текущей комнаты.</para>
+/// </summary>
+public sealed class UITabletRoomContentSupplier : MonoBehaviour
+{
+    [SerializeField] private GameObject _content;
+
+    /// <summary>
+    /// <para>Получить контент текущей комнаты.</para>
+    /// </summary>
+    /// <returns>контент текущей комнаты</returns>
+    public GameObject GetRoomContent()
+    {
+        return _content;
+    }
+}
