--- Assets/ApplicationContent/Player/Tablet/PC/Scripts/UI/UITabletRoomContentSupplier.cs
@@ -1,21 +1,0 @@
-using UnityEngine;
-
-namespace Player.Tablet.PC.UI
-{
-    /// <summary>
-    /// <para>Компонент передающий планшету контент текущей комнаты.</para>
-    /// </summary>
-    public sealed class UITabletRoomContentSupplier : MonoBehaviour
-    {
-       [SerializeField] private GameObject _content;
-
-       /// <summary>
-       /// <para>Получить контент текущей комнаты.</para>
-       /// </summary>
-       /// <returns>контент текущей комнаты</returns>
-       public GameObject GetRoomContent()
-       {
-           return _content;
-       }
-    }
-}
